Split off Ledger.SupportsProtocol
