+++ ouroboros-consensus/src/Ouroboros/Consensus/Ledger/Extended.hs
@@ -34,6 +34,7 @@ import           Ouroboros.Network.Block (HeaderHash)
 import           Ouroboros.Consensus.Block
 import           Ouroboros.Consensus.HeaderValidation
 import           Ouroboros.Consensus.Ledger.Abstract
+import           Ouroboros.Consensus.Ledger.SupportsProtocol
 import           Ouroboros.Consensus.Protocol.Abstract
 import           Ouroboros.Consensus.Util (repeatedlyM)
