+++ ouroboros-consensus/src/Ouroboros/Consensus/Node/Run.hs
@@ -30,6 +30,7 @@ import           Ouroboros.Consensus.BlockchainTime (SystemStart)
 import           Ouroboros.Consensus.HeaderValidation
 import           Ouroboros.Consensus.Ledger.Abstract
 import           Ouroboros.Consensus.Ledger.Extended
+import           Ouroboros.Consensus.Ledger.SupportsProtocol
 import           Ouroboros.Consensus.Mempool
 import           Ouroboros.Consensus.Node.Exit (ExitReason)
 import           Ouroboros.Consensus.Node.NetworkProtocolVersion
