+++ ouroboros-consensus/src/Ouroboros/Consensus/NodeKernel.hs
@@ -57,6 +57,7 @@ import           Ouroboros.Consensus.ChainSyncClient
 import           Ouroboros.Consensus.HeaderValidation
 import           Ouroboros.Consensus.Ledger.Abstract
 import           Ouroboros.Consensus.Ledger.Extended
+import           Ouroboros.Consensus.Ledger.SupportsProtocol
 import           Ouroboros.Consensus.Mempool
 import           Ouroboros.Consensus.Mempool.TxSeq (TicketNo)
 import           Ouroboros.Consensus.Node.Run (RunNode (..))
